- changes to soapui project - moved errordetails to baseclass
