+++ src/main/java/nl/haarlem/translations/zdstozgw/requesthandler/impl/BasicRequestHandler.java
@@ -6,6 +6,7 @@ import nl.haarlem.translations.zdstozgw.converter.ConverterException;
 import nl.haarlem.translations.zdstozgw.requesthandler.RequestHandler;
 import nl.haarlem.translations.zdstozgw.translation.zds.model.ZdsBv03;
 import nl.haarlem.translations.zdstozgw.translation.zds.model.ZdsFo03;
+import nl.haarlem.translations.zdstozgw.translation.zds.model.ZdsStuurgegevens;
 import nl.haarlem.translations.zdstozgw.translation.zds.model.ZdsZakLk01;
 import nl.haarlem.translations.zdstozgw.utils.XmlUtils;
 
@@ -42,27 +43,9 @@ public class BasicRequestHandler extends RequestHandler {
 			return new ResponseEntity<>(response, HttpStatus.OK);	        
 		}
 		catch(Exception ex) {
-			log.warn("request for path: /" + path + "/ with soapaction: " + soapAction, ex);
-			
-			// get the stacktrace
-			var swriter = new java.io.StringWriter();
-			var pwriter = new java.io.PrintWriter(swriter);
-			ex.printStackTrace(pwriter);
-			var stacktrace = swriter.toString();			
-			 
-	        var fo03 = new ZdsFo03(zdsRequest.stuurgegevens);
-	        fo03.body = new ZdsFo03.Body();
-	        https://www.gemmaonline.nl/images/gemmaonline/4/4f/Stuf0301_-_ONV0347_%28zonder_renvooi%29.pdf
-	        fo03.body.code = "StUF058";
-	        fo03.body.plek = "server";
-	        fo03.body.omschrijving = ex.toString();
-	        fo03.body.entiteittype = "";
-	        fo03.body.details = stacktrace;
-	        fo03.body.detailsXML = request;                    
-	        
+			var fo03 = getErrorZdsDocument(ex,path, soapAction, request, zdsRequest.stuurgegevens);
 	        var response = XmlUtils.getSOAPFaultMessageFromObject(SOAPConstants.SOAP_RECEIVER_FAULT, ex.toString(), fo03);
-	        
-	        return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+	        return new ResponseEntity<>(response, ex instanceof ConverterException ? ((ConverterException) ex).getHttpStatus() : HttpStatus.INTERNAL_SERVER_ERROR);
 		}
     }
 }
